Update README.md
u may be shit at blender so heres a little tut. so ofc you need blender (idk how to spell so: i recomend blender 4.3.2 or an 4. somethn) and when you done, go to render and render image and when it done rendering. go to image and save image and go to where it is and ten save or somethn so yeah by (if you need any help because your bad at blender or a fuckin idiot then dm me on discord)